refactor(rank): 게임별로 랭킹을 조회하도록 API 스펙을 변경한다 (#148)
