+++ src/main/java/com/snackgame/server/rank/domain/BestScore.java
@@ -1,4 +1,4 @@
-package com.snackgame.server.rank.applegame.domain;
+package com.snackgame.server.rank.domain;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
@@ -18,7 +18,7 @@ import lombok.NoArgsConstructor;
 @Getter
 @Table(uniqueConstraints = @UniqueConstraint(
         name = "unique_best_score_in_a_season",
-        columnNames = {"season_id", "owner_id"}
+        columnNames = {"seasonId", "ownerId"}
 ))
 @Entity
 public class BestScore {
@@ -29,9 +29,9 @@ public class BestScore {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
     private int score = 0;
-    @Column(name = "owner_id", nullable = false)
+    @Column(nullable = false)
     private Long ownerId;
-    @Column(name = "season_id", nullable = false)
+    @Column(nullable = false)
     private Long seasonId;
     private Long sessionId = null;
