+++ src/main/java/com/snackgame/server/rank/domain/Season.java
@@ -1,4 +1,4 @@
-package com.snackgame.server.rank.applegame.domain;
+package com.snackgame.server.rank.domain;
 
 import java.time.LocalDateTime;
