+++ src/main/java/com/snackgame/server/rank/domain/SeasonRepository.java
@@ -1,4 +1,4 @@
-package com.snackgame.server.rank.applegame.domain;
+package com.snackgame.server.rank.domain;
 
 import java.util.Optional;
